+++ src/app/Pages/BotPlanner/DataTable.tsx
@@ -4,6 +4,7 @@ import { Delete as DeleteIcon } from '@material-ui/icons';
 import styled from 'styled-components'
 
 import { useGlobalData } from '@/app/Context/DataContext';
+import { storageItem } from '@/app/Features/LocalStorage/LocalStorage';
 
 import { parseNumber } from '@/utils/number_formatting';
 import {
@@ -155,13 +156,14 @@ interface Type_DataTable {
   updateLocalBotData: any
 }
 const DataTable = ({ localBotData, updateLocalBotData }: Type_DataTable) => {
+  const localStorageSortName = storageItem.tables.BotPlanner.sort;
 
   const state = useGlobalData()
 
   const { data: { metricsData: { totalBankroll } } } = state;
 
 
-  // handling this locally becauase it does not need to be saved yet.
+  // handling this locally because it does not need to be saved yet.
   const handleOnOff = (e: any) => {
     updateLocalBotData((prevState: Type_Query_bots[]) => {
       const newRows = prevState.map((row: Type_Query_bots) => {
@@ -386,6 +388,7 @@ const DataTable = ({ localBotData, updateLocalBotData }: Type_DataTable) => {
             // autoResetPage={false}
             manualSortBy={true}
             updateLocalBotData={handleEditCellChangeCommitted}
+            localStorageSortName={localStorageSortName}
             //@ts-ignore
             getHeaderProps={column => ({